Add files via upload
Created the classes Transaction (from Project4), TransactionChecker (Observable), TransactionLogger.

Creating the TransactionChecker class:
Made this class a derived class that extends the base class Observable. 
In the while loop, created a Transaction object. For the transaction id, used an autogenerated id that starts with 1001 and increases by 1 for each new Transaction object created.
After the Transaction object is check it runs a while loop,  an if statement is executed, that notifies the observers if a transaction was found if amount is greater than $50,000 and the buyer is Eugene Eko. It Notifies all observers by sending them the Transaction object. To do this, I Observable methods setChanged and notifyObservers.

Creating the TransactionLogger class:
It implements the Observer interface, which means that an update method was required. public void update(Observable obs, Object info)
Wrote the info object (which is actually a Transaction object) out to the log file. After that, it flushes the output buffer, which forces the info to be written to the disk. The @Override statement, indicates that you are either overriding a method from a base class or implementing a required method from an interface. 

Creating a TestTransactionChecker class. In a main method:
Declared and instantiated a new TransactionLogger object. 
Added the TransactionLogger object as an observer to the TransactionChecker object.