Add files via upload
Implements and test the base class Transaction, and the derived classes GoldTransaction, and LumberTransaction.

The Transaction class implements the PurchaseReceipt interface, which inherits from the Comparable interface. This means that in addition to supplying the methods getId, getBuyer, getSeller, and getAmount, a compareTo method is also supplied. It returns 1 if this._id > other._id, it should return -1 if this._id > other._id; or return 0 if this._id == other._id.

Included a traditional test class Test1 and a JUnit test class Test2 that tests all of the methods in the Transaction class, and all of the non-inherited methods in the derived classes. Overridden methods are tested again for the derived classes. 

Implemented and tested a TransactionManager class that manages objects that implement the PurchaseReceipt interface.  These objects are Transaction objects and the objects of their derived classes.

The Transaction and TransactionManager classes implement the Serializable interface so that it can be serialized to or deserialized from a .ser file. 

Wrote a traditional test class Test3 that tests the TransactionManager class.