Add files via upload
Print Address with POSTNET bar code. (POSTNET means Postal Numeric Encoding Technique.)

For each line the address is written like the examples to the output file labels.txt:
Anna Lee
123 Nice Street
Memphis TN 38141-8346
|::||:|::|::::||:|::|:::|||::|:::||::|::|:||::::|:||

Stan Smith
456 De La Vina Street
Santa Barbara CA 93101-3298
||:|::::||::::||||::::::||::||:::|:||:|::|::|::|::||

The bar code is encoded using this table:
 
1	2	3	4	5	6	7	8	9	0
:::||	::|:|	::||:	:|::|	:|:|:	:||::	|:::|	|::|:	|:|::	||:::

In addition to encoding the 9 digits of the zip code, the bar code includes initial (i) and terminal (t) frame bars and a check sum (cs), which is 10 - the sum of the zip code digits mod 10. Thus the zip code 38141-8346 is encoded like this:
i   3     8     1     4     1     8     3     4     6     cs  t
| ::||: |::|: :::|| :|::| :::|| |::|: ::||: :|::| :||:: ::|:| |

The sum of the digits is 38 and 38 % 10 is 8, so the  sum is 10 - 8 = 2. The initial and terminal frame bars are always |.